corpai-backend/requirements.txt
# CorpAI Backend — Dependências Python
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
chromadb==0.6.3
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-json-logger==3.2.1
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
sse-starlette==2.2.1
alembic==1.14.1
